MeiDuo/requirements.txt
itsdangerous==1.1.0
drf_extensions==0.4.0
fdfs_client_py==1.2.6
celery==4.3.0
django_ckeditor==5.7.0
djangorestframework_jwt==1.11.0
redis==3.2.1
xadmin==0.6.1
django_redis==4.10.0
Django==1.11.11
drf_haystack==1.8.4
PyMySQL==0.9.3
Pillow==6.0.0
fdfs_client==4.0.7
django_haystack==2.8.1
djangorestframework==3.9.4
